Revert "Merge branch 'main' into Jonathan_UI"
This reverts commit 3f9f8f5a7cbd0826c9622563680e9318f981e7c0, reversing
changes made to 3fa31ddbfba6567159c98fd47c4bfee01f3ff092.

+++ Viking Attack/Assets/Scripts/Character_scripts/CharacterBase.cs
@@ -19,7 +19,7 @@ namespace ItemNamespace
         [SerializeField] private float attackCooldown;
         [SerializeField] private int damage;
         [SerializeField] private float chasingSpeedMultiplier;
-        [SerializeField] private float moveSpeed;
+        [SerializeField] private int moveSpeed;
         [SerializeField] private float maxHealth;
         [SerializeField] private float experienceRadius;
         [SerializeField] private float experience;
@@ -65,7 +65,7 @@ namespace ItemNamespace
             return chasingSpeedMultiplier;
         }
 
-        public float GetMovementSpeed()
+        public int GetMovementSpeed()
         {
             return moveSpeed;
         }
